+++ tests/test_formula.py
@@ -1,7 +1,7 @@
-import pytest
 import re
-
 from typing import Any, Callable, Optional
+
+import pytest
 
 from snuba_sdk.column import Column
 from snuba_sdk.conditions import Condition, Op
@@ -16,27 +16,55 @@ tests = [
         id="basic formula test",
     ),
     pytest.param(
-        formula(ArithmeticOperator.PLUS, [Timeseries(metric=Metric(public_name="foo"), aggregate="sum"), 1], None, None),
+        formula(
+            ArithmeticOperator.PLUS,
+            [Timeseries(metric=Metric(public_name="foo"), aggregate="sum"), 1],
+            None,
+            None,
+        ),
         None,
         id="timeseries and number formula test",
     ),
     pytest.param(
-        formula(ArithmeticOperator.PLUS, [Timeseries(metric=Metric(public_name="foo"), aggregate="sum"), 1], [Condition(Column("tags[transaction]"), Op.EQ, "foo")], None),
+        formula(
+            ArithmeticOperator.PLUS,
+            [Timeseries(metric=Metric(public_name="foo"), aggregate="sum"), 1],
+            [Condition(Column("tags[transaction]"), Op.EQ, "foo")],
+            None,
+        ),
         None,
         id="filters in formula",
     ),
     pytest.param(
-        formula(ArithmeticOperator.PLUS, [Timeseries(metric=Metric(public_name="foo"), aggregate="sum"), 1], None, [Column("tags[status_code]")]),
+        formula(
+            ArithmeticOperator.PLUS,
+            [Timeseries(metric=Metric(public_name="foo"), aggregate="sum"), 1],
+            None,
+            [Column("tags[status_code]")],
+        ),
         None,
         id="groupby in formula",
     ),
     pytest.param(
-        formula(ArithmeticOperator.PLUS, [Timeseries(metric=Metric(public_name="foo"), aggregate="sum"), Timeseries(metric=Metric(public_name="bar"), aggregate="sum")], None, None),
+        formula(
+            ArithmeticOperator.PLUS,
+            [
+                Timeseries(metric=Metric(public_name="foo"), aggregate="sum"),
+                Timeseries(metric=Metric(public_name="bar"), aggregate="sum"),
+            ],
+            None,
+            None,
+        ),
         None,
         id="timeseries in formula",
     ),
     pytest.param(
-        formula(42, [Timeseries(metric=Metric(public_name="foo"), aggregate="sum"), 1], None, None),
+        formula(
+            42,
+            [Timeseries(metric=Metric(public_name="foo"), aggregate="sum"), 1],
+            None,
+            None,
+        ),
         InvalidFormulaError("formula '42' must be a ArithmeticOperator"),
         id="invalid operator type",
     ),
@@ -46,7 +74,12 @@ tests = [
         id="invalid parameters",
     ),
     pytest.param(
-        formula(ArithmeticOperator.MULTIPLY, [Timeseries(metric=Metric(public_name="foo"), aggregate="sum"), "foo"], None, None),
+        formula(
+            ArithmeticOperator.MULTIPLY,
+            [Timeseries(metric=Metric(public_name="foo"), aggregate="sum"), "foo"],
+            None,
+            None,
+        ),
         InvalidFormulaError("parameter 'foo' of formula multiply is an invalid type"),
         id="unsupported parameter for operator",
     ),